Inline bundle static text during snapshotting
