+++ cli/js/compiler.ts
@@ -6,9 +6,9 @@ import "./globals.ts";
 import "./ts_global.d.ts";
 
 import { TranspileOnlyResult } from "./compiler_api.ts";
+import { oldProgram } from "./compiler_bootstrap.ts";
 import { setRootExports } from "./compiler_bundler.ts";
 import {
-  ASSETS,
   defaultBundlerOptions,
   defaultRuntimeCompileOptions,
   defaultTranspileOptions,
@@ -27,9 +27,7 @@ import {
   WriteFileState,
   processConfigureResponse
 } from "./compiler_util.ts";
-import { core } from "./core.ts";
 import { Diagnostic } from "./diagnostics.ts";
-import * as dispatch from "./dispatch.ts";
 import { fromTypeScriptDiagnostic } from "./diagnostics_util.ts";
 import * as os from "./os.ts";
 import { assert } from "./util.ts";
@@ -73,26 +71,6 @@ interface CompileResult {
   emitSkipped: boolean;
   diagnostics?: Diagnostic;
 }
-
-// this is used to generate the foundational AST for all other compilations, so
-// it can be cached as part of the
-let oldProgram: ts.Program;
-
-((): void => {
-  const ops = core.ops();
-  for (const [name, opId] of Object.entries(ops)) {
-    const opName = `OP_${name.toUpperCase()}`;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (dispatch as any)[opName] = opId;
-  }
-  const host = new Host({ writeFile(): void {} });
-  const options = host.getCompilationSettings();
-  oldProgram = ts.createProgram({
-    rootNames: [`${ASSETS}/bootstrap.ts`],
-    options,
-    host
-  });
-})();
 
 // bootstrap the runtime environment, this gets called as the isolate is setup
 self.denoMain = function denoMain(compilerType?: string): void {
